fix(#239): Delete game with existing plays (#240)
* fix(#239): Delete game with existing plays

- Fix Navbar for attendee-check-in (small fix)
- Add two tests for deleting play to win and library
- Fixed multiple issues with deleting play to win games
- Fixed library games removing foreign key constraints

closes: #239

* update to next 15.0.x

* forgot to add package.json

* upgrade to next 15 and fix key error

* upgrade to 15.0.x in nextjs

* add not-found.js per the next documentation

* upgrade to tsx instead of js

* add error.tsx

* downgrade swagger

+++ apps/admin/app/api/library/[id]/route.ts
@@ -36,9 +36,11 @@ export const dynamic = "force-dynamic"
  */
 export async function GET(
   request: NextRequest,
-  { params }: { params: { id: string } }
+  { params }: { params: Promise<{ id: string }> }
 ) {
-  const response = await GetLibraryItemById(Number(params.id))
+  const libraryItemId = Number((await params).id)
+
+  const response = await GetLibraryItemById(libraryItemId)
   if (response === null) NextResponse.json({ message: "Game not found" }, { status: 404 })
 
   return NextResponse.json<ILibraryItem>(response ?? {} as ILibraryItem, { status: 200 })
@@ -77,15 +79,30 @@ export async function GET(
  *             schema:
  *                message: string
  */
-export async function DELETE(request: NextRequest, { params }: { params: { id: string } }) {
+export async function DELETE(request: NextRequest, { params }: { params: Promise<{ id: string }> }) {
+  const libraryItemId = (await params).id
+  const libraryItem = await prisma.libraryItem.findFirst({
+    where: { id: Number(libraryItemId)}
+  })
+
+  if (libraryItem === null || libraryItem === undefined)
+    return NextResponse.json({ message: `Library item with ${libraryItemId} not found` }, { status: 404 });
+
+
+  await prisma.libraryCheckoutEvent.deleteMany({
+    where: {
+      libraryCopyId: Number(libraryItemId)
+    }
+  })
+
   await prisma.libraryItem.delete({
-    where: { id: Number(params.id) },
+    where: { id: Number(libraryItemId) },
   });
 
   await prisma.centralizedBarcode.delete({
     where: {
       entityType_entityId: {
-        entityId: Number(params.id),
+        entityId: Number(libraryItemId),
         entityType: "LibraryItem",
       },
     },
@@ -93,7 +110,7 @@ export async function DELETE(request: NextRequest, { params }: { params: { id: s
 
   return NextResponse.json(
     {
-      message: `Successfully deleted library item - ${params.id}`,
+      message: `Successfully deleted library item - ${libraryItemId}`,
     },
     { status: 200 }
   );
